requirements.txt
streamlit==1.40.0
streamlit-drawable-canvas
easyocr
opencv-python-headless
numpy
Pillow
